+++ inc/Engine/Admin/PerformanceMonitoring/Settings/Controller.php
@@ -1,18 +1,12 @@
 <?php
 declare(strict_types=1);
 
-namespace WP_Rocket\Engine\Admin\PerformanceMonitoring;
+namespace WP_Rocket\Engine\Admin\PerformanceMonitoring\Settings;
 
 use WP_Rocket\Abstract_Render;
 use WP_Rocket\Engine\License\API\User;
-use WP_Rocket\Event_Management\Subscriber_Interface;
 
-/**
- * Handle Add-On license status display
- *
- * @since 3.20
- */
-class SettingsSubscriber extends Abstract_Render implements Subscriber_Interface {
+class Controller extends Abstract_Render {
 	/**
 	 * User API client
 	 *
@@ -32,17 +26,6 @@ class SettingsSubscriber extends Abstract_Render implements Subscriber_Interface
 	}
 
 	/**
-	 * Events this subscriber listens to
-	 *
-	 * @return array
-	 */
-	public static function get_subscribed_events() {
-		return [
-			'rocket_dashboard_after_account_data' => [ 'display_addon_status', 9 ], // Higher priority than RocketCDN.
-		];
-	}
-
-	/**
 	 * Displays the Add-On license status on the dashboard tab
 	 *
 	 * @since 3.20
@@ -50,7 +33,6 @@ class SettingsSubscriber extends Abstract_Render implements Subscriber_Interface
 	 * @return void
 	 */
 	public function display_addon_status() {
-
 		if ( (bool) rocket_get_constant( 'WP_ROCKET_WHITE_LABEL_ACCOUNT' ) ) {
 			return;
 		}
@@ -60,10 +42,8 @@ class SettingsSubscriber extends Abstract_Render implements Subscriber_Interface
 		$status_text  = __( 'No Subscription', 'rocket' );
 		$service_name = __( 'Rocket Insights', 'rocket' );
 		$sku          = $this->user->get_pma_addon_sku_active();
-
 		$upgrade_skus = $this->user->get_pma_addon_upgrade_skus( $sku );
-
-		$is_active = $this->user->is_pma_addon_active( $sku );
+		$is_active    = $this->user->is_pma_addon_active( $sku );
 
 		if ( $is_active ) {
 			$label        = __( 'Next Billing Date', 'rocket' );
